+++ lib/Xmu/StrToLong.c
@@ -21,7 +21,7 @@ used in advertising or otherwise to promote the sale, use or other dealings
 in this Software without prior written authorization from The Open Group.
 
 */
-/* $XFree86: xc/lib/Xmu/StrToLong.c,v 1.4 1998/08/20 13:59:41 dawes Exp $ */
+/* $XFree86: xc/lib/Xmu/StrToLong.c,v 1.5 1998/10/03 09:06:35 dawes Exp $ */
 
 #include <stdio.h>
 #include <X11/Intrinsic.h>
@@ -51,7 +51,7 @@ XmuCvtLongToString(Display *dpy, XrmValuePtr args, Cardinal *num_args,
 		   XrmValuePtr fromVal, XrmValuePtr toVal, XtPointer *data)
 {
   static char buffer[32];
-  int size;
+  size_t size;
 
   if (*num_args != 0)
     XtWarning("Long to String conversion needs no extra arguments");
